requirements.txt
transformers==4.18.0
torch==2.3.0
torchaudio==2.3.0
librosa==0.9.1
note-seq==0.0.3
pretty-midi==0.2.9
einops==0.4.1
hydra-core==1.2.0
pytorch-lightning==2.2.3
wandb==0.16.6
immutabledict==4.2.0
timm==0.9.16
matplotlib==3.8.4
mir_eval==0.7
huggingface_hub==0.23.0
numpy==1.26.4
scipy==1.13.0
soundfile==0.12.1
tqdm==4.66.2
PyYAML==6.0.1
omegaconf==2.2.3
absl-py==2.1.0
